Ignore Disposables created on return statements
When an IDisposable object is created on the return statement
of a method, the DisposableVariableNotDisposedAnalyzer creates a
Diagnostic saying that the said object should be disposed.
Trying to Fix Occurences of said Diagnostic causes VS to crash
when CreateUsing (class DisposableVariableNotDisposedCodeFixProvider) is
called. The crash is caused by an InvalidCastException thrown by
CreateUsing on this cast:

(ExpressionStatementSyntax)objectCreation.Parent.

Since the disposable object is being created and returned, the Parent
property of objectCreation is actually of type ReturnStatementSyntax,
causing the exception to be thrown.

I have fixed the DisposableVariableNotDisposedAnalyzer so it ignores nodes
whose Parents are of type ReturnStatementSyntax.