Reorganize StringFormatArgsAnalyzer to use the correct severities
Fixes #585
StringFormatArgsAnalyzer was always raising an error, when it is not an
error to have extra parameters. So we have taken an extra id, CC0111 to
use as the warning, when string.Format is being provided more arguments
than the format string is requiring.
This will now raise a warning:

    string.Format("{0}", "a", "b");

When string.Format has less arguments than the format string requires,
we are still raising CC0056, as an error, as it will generate a
runtime error.
This will now raise an error:

    string.Format("{0}{1}", "a");

Other verifications the analyzer was already checking continue as
before.