Make CodeFixVerifier to always check for new diags
Before, if a codeFixIndex was supplied, CodeFixVerifier
was not verifying if the analyzer had raised new
diagnostics. Now this is fixed.
When this was enabled, several code fix providers
started generating test errors. This commit also fixes
all of those.
Also, discovered that the
DisposablesShouldCallSuppressFinalize is assuming
System is imported, added a skipped test to fix that.
It was also not working correctly with explicitly
implemented interfaces, and this is now fixed, also
the tests were fixed.
Also, fixed ParameterRefactory which was generating
correct code in text, but an incorrect syntax tree.