Update package names to solve error 500
This the error we were getting:
Failed to process request. 'The title of your package, 'CodeCracker', is similar to the ID of an existing package, which can cause confusion with our users. Please modify the title of your package and try uploading again.'.
The remote server returned an error: (500) Internal Server Error..