Fix bug for CC0118 Remove to string
The string is only removed if safe to do.

Fixes #866.

Implemented option 2 as requested in #866.
This fix now looks up the types of both sides
of the plus operator and tries removes those
candidates, that might break things.

It turned out, that some of the checks in
CheckAddOperationOverloadsOfTypes are not
needed (After the second check every other check
returns false). Essentially right now the removal of
"ToString" is only save if either the underlying type
is a string (as in 1 + "a".ToString()) or the type on
the other side of the plus operator is a string (as in
"a" + 1.ToString()). I kept all the other checks for
future enhancements to make clear that there
are several cases to consider. Further I added lots
of unit tests to cover these cases. The idea is that
CheckAddOperationOverloadsOfTypes can easily
be expanded in the future to support more cases
without breaking anything. Those redundant checks
are also very cheap and are essentially just simple
comparisons.